adk_agent/mentor_scout/requirements.txt
google-adk
google-cloud-bigquery
pandas
python-dotenv
tavily-python
